+++ Mutagen.Bethesda.Fallout4/Records/Major Records/Perk.cs
@@ -567,7 +567,7 @@ partial class PerkBinaryOverlay
 
     partial void EffectsCustomParse(
         OverlayStream stream,
-        long finalPos,
+        int finalPos,
         int offset,
         RecordType type,
         PreviousParse lastParsed)
