+++ Mutagen.Bethesda.Fallout4/Records/Major Records/Quest.cs
@@ -335,7 +335,7 @@ partial class QuestBinaryOverlay
 
     public IReadOnlyList<IDialogBranchGetter> DialogBranches { get; private set; } = Array.Empty<IDialogBranchGetter>();
 
-    partial void DialogConditionsCustomParse(OverlayStream stream, long finalPos, int offset, RecordType type, PreviousParse lastParsed)
+    partial void DialogConditionsCustomParse(OverlayStream stream, int finalPos, int offset, RecordType type, PreviousParse lastParsed)
     {
         DialogConditions = ConditionBinaryOverlay.ConstructBinayOverlayList(stream, _package);
     }
